Improve exit flame particles

+++ Assets/Scripts/Events/EventData/LivesChangeData.cs
@@ -3,7 +3,7 @@ public class LivesChangeData
     public int change;
     public int remaining;
     public int starting;
-    public float percent => starting > 0 ? remaining / starting : 0;
+    public float percent => starting > 0 ? remaining / (float)starting : 0;
 
     public LivesChangeData(int remaining, int change, int starting)
     {
